files

+++ ft_strlen.c
@@ -10,10 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int			ft_strlen(const char *s)
+#include "libft.h"
+
+size_t			ft_strlen(const char *s)
 {
-	int c;
-	
+	size_t c;
+
 	c = 0;
 	while (s[c])
 		c++;
